Added debounce to chart resizing.
Debounce prevents the function attached to the event handler from being executed until a specified time period after the last time it gets called. That way the chart only gets re-rendered far less frequently when the browser is resized. But it will always get re-rendered after the browser is done resizing. This can be a huge performance increase in some circumstances.